pyproject.toml
[tool.poetry]
name = "fastmedapp"
version = "0.1.0"
description = "A fast medical support application implemented using django and angular"
authors = ["BloodCharry <vladvorobei1994@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
psycopg2 = "^2.9.9"
djoser = "^2.2.3"
simplejwt = "^2.0.1"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
packaging = "^24.2"
django-cors-headers = "^4.6.0"
django-jazzmin = "^3.0.1"
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.154.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
